requirements.txt
# AutoDJ Development Dependencies

# Code Formatting & Linting
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.2

# Spotify Integration
spotipy==2.25.1
python-dotenv==1.0.0
pydantic == 2.11

# YouTube Download
yt-dlp==2025.9.26

# OpenAI Integration
openai==1.49.0

# Development Tools
autopep8==2.0.4
pylint==3.0.3

